Extracted the git functions into their own module for easier reuse
